Add authorization when creating/deleting/editing artworks and importing data

+++ backend/controllers/artworks.ts
@@ -5,7 +5,7 @@ const asyncWrapper = require("../middleware/async")
 const Artwork = require("../models/artwork")
 const Collection = require("../models/collection")
 const Category = require("../models/category")
-
+const jwt = require("jsonwebtoken")
 const ObjectId = require("mongodb").ObjectId
 
 const getAllArtworks = async (req: Request, res: Response, next: NextFunction) => {
@@ -136,8 +136,15 @@ const filterArtworks = async (req: Request, res: Response, next: NextFunction): 
 
 const createArtwork = (async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const newArtwork = await Artwork.create(req.body)
-        return res.status(201).json(newArtwork)
+        const token = req.headers.authorization?.split(" ")[1]
+        if (!token) return res.status(401).json({ error: 'Access denied' });
+        try {
+            const decoded = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET as string)
+            const newArtwork = await Artwork.create(req.body)
+            return res.status(201).json(newArtwork)
+        } catch (error) {
+            return res.status(401).json({ error: 'Access denied' });
+        }
     } catch (error) {
         next(error)
     }
@@ -145,8 +152,15 @@ const createArtwork = (async (req: Request, res: Response, next: NextFunction) =
 
 const editArtwork = (async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const editedArtwork = await Artwork.replaceOne({_id: req.params.artworkId}, req.body)
-        return res.status(201).json(editedArtwork)
+        const token = req.headers.authorization?.split(" ")[1]
+        if (!token) return res.status(401).json({ error: 'Access denied' });
+        try {
+            const decoded = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET as string)
+            const editedArtwork = await Artwork.replaceOne({_id: req.params.artworkId}, req.body)
+            return res.status(201).json(editedArtwork)
+        } catch (error) {
+            return res.status(401).json({ error: 'Access denied' });
+        }
     } catch (error) {
         next(error)
     }
@@ -154,7 +168,6 @@ const editArtwork = (async (req: Request, res: Response, next: NextFunction) => 
 
 const deleteArtwork = asyncWrapper(async (req: Request, res: Response, next: NextFunction) => {
     const artworkId = req.params.artworkId
-
     try {
         if (!mongoose.isValidObjectId(artworkId)) {
             return res.status(400).json("Invalid artwork id")
@@ -174,21 +187,28 @@ const deleteArtwork = asyncWrapper(async (req: Request, res: Response, next: Nex
 
 const batchDeleteArtworks = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const artworksToDelete = req.params.artwork
+        const token = req.headers.authorization?.split(" ")[1]
+        if (!token) return res.status(401).json({ error: 'Access denied' });
+        try {
+            const decoded = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET as string)
+            const artworksToDelete = req.params.artwork
+            if (!artworksToDelete) {
+                return res.status(400).send({ message: "Artworks not found" })
+            }
+            const artworksToDeleteList = artworksToDelete.split(",")
+            const existingArtworks = await Artwork.find({ _id: { $in: artworksToDeleteList } })
 
-        if (!artworksToDelete) {
-            return res.status(400).send({ message: "Artworks not found" })
+            if (existingArtworks.length === 0) {
+                return res.status(404).send({ message: "Artworks not found" })
+            }
+
+            const result = await Artwork.deleteMany({ _id: { $in: artworksToDeleteList } })
+
+            res.status(200).json({ message: req.params.artwork, deletedCount: result.deletedCount })
+        } catch (error) {
+            return res.status(401).json({ error: 'Access denied' });
         }
-        const artworksToDeleteList = artworksToDelete.split(",")
-        const existingArtworks = await Artwork.find({ _id: { $in: artworksToDeleteList } })
-
-        if (existingArtworks.length === 0) {
-            return res.status(404).send({ message: "Artworks not found" })
-        }
-
-        const result = await Artwork.deleteMany({ _id: { $in: artworksToDeleteList } })
-
-        res.status(200).json({ message: req.params.artwork, deletedCount: result.deletedCount })
+        
     } catch (error) {
         next(error)
     }
